requirements/base.txt
Django==3.2.13
Pillow==9.1.1
psycopg2-binary==2.9.3
djangorestframework==3.12.0
djangorestframework-simplejwt==4.6.0
PyJWT==2.4.0
djoser==2.0.3
